requirements.txt
# Core LiveKit Agent SDK
livekit-agents[google] ~= 1.0

# LiveKit plugins
livekit-plugins-deepgram>=0.6.0
livekit-plugins-cartesia>=0.6.0
livekit-plugins-noise-cancellation ~= 0.2

# Environment variables
python-dotenv>=1.0.0